requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.44.0
networkx==3.3
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.46.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
